# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Webshop API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/customer/{customer_id}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "description": "This is the route for getting a customer",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/customer": {
      "post": {
        "tags": [
          "Customer"
        ],
        "description": "This is the route for getting creating a customer",
        "parameters": [
          {
            "name": "customer",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddCustomer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/customers": {
      "get": {
        "tags": [
          "Customer"
        ],
        "description": "This is the route for getting all customers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login/user": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "This is the route for the user login. this is needed for getting access to certain routes",
        "parameters": [
          {
            "name": "login",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login/employee": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "This is the route for the employee login. this is needed for getting access to certain routes",
        "parameters": [
          {
            "name": "login",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/LoginEmployee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/register/user": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "This is the route for the creating a new user",
        "parameters": [
          {
            "name": "newUser",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/RegisterUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "Login"
        ],
        "description": "This is the route for the user/employee to logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/hashedpassword": {
      "get": {
        "tags": [
          "Login"
        ],
        "description": "",
        "parameters": [
          {
            "name": "pw",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/order": {
      "post": {
        "tags": [
          "Order"
        ],
        "description": "This is the route for the creating a new order. a bit bugged since i dont know how to work with multiple schemas you have to copy/paste response 200 while being logged in to checkout user buying something. Other than that for some reason the product array only uses the last index",
        "parameters": [
          {
            "name": "key",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/OrderTypeTwo"
            },
            "description": "This is for testing where a non user sends to a PO box"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/OrderTypeOne"
            },
            "description": "This is for testing where user sends to own address you need to be logged in as a user for this"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "Not Implemented"
          }
        }
      }
    },
    "/order/{order_id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "This is the route for getting a specific order.",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/{page}/{size}": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "This is the route for getting all orders. can use pagination. use 0 value if wanna use default values of p1/s1000",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "size",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/search/{key}/{value}/{page}": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "This is the route for seaching in orders. with pagination. key, value EX. KEY: order_status VALUE: shipped ",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orderoverview/user/{user_id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "This is the route for getting a users order overview.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddCustomer": {
      "type": "object",
      "properties": {
        "customer": {
          "type": "object",
          "properties": {
            "first_name": {
              "type": "string",
              "example": "David"
            },
            "last_name": {
              "type": "string",
              "example": "Clements"
            },
            "street": {
              "type": "string",
              "example": "59161 Chandler Rapids"
            },
            "email": {
              "type": "string",
              "example": "nguyenpatrick@gmail.com"
            },
            "phone": {
              "type": "string",
              "example": "84081093"
            },
            "cities_postal_code": {
              "type": "number",
              "example": 9000
            },
            "countries_iso": {
              "type": "string",
              "example": "SE"
            },
            "users_user_id": {
              "type": "number",
              "example": 15
            },
            "is_user_profile": {
              "type": "number",
              "example": 0
            }
          }
        }
      }
    },
...
    "LoginUser": {
      "type": "object",
      "properties": {
        "login": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string",
              "example": "user1"
            },
            "password": {
              "type": "string",
              "example": "user1kode"
            }
          }
        }
      }
    },
    "LoginEmployee": {
      "type": "object",
      "properties": {
        "login": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "example": "admin@gmail.com"
            },
            "password": {
              "type": "string",
              "example": "adminkode"
            }
          }
        }
      }
    },
    "RegisterUser": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string",
              "example": "oliv3f342"
            },
            "password": {
              "type": "string",
              "example": "madoli123"
            },
            "created_date": {
              "type": "string",
              "example": "2011-07-29"
            },
            "is_archived": {
              "type": "number",
              "example": 0
            },
            "last_logged_in": {
              "type": "string",
              "example": "2006-05-29"
            },
            "customers": {
              "type": "object",
              "properties": {
                "first_name": {
                  "type": "string",
                  "example": "Jennifer"
                },
                "last_name": {
                  "type": "string",
                  "example": "Chandler"
                },
                "street": {
                  "type": "string",
                  "example": "8650 Michael Grove"
                },
                "email": {
                  "type": "string",
                  "example": "gjoseph@gmail.com"
                },
                "phone": {
                  "type": "string",
                  "example": "40949002"
                },
                "cities_postal_code": {
                  "type": "number",
                  "example": 1000
                },
                "countries_iso": {
                  "type": "string",
                  "example": "DK"
                },
                "users_user_id": {
                  "type": "number",
                  "example": 123
                },
                "is_user_profile": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        }
      }
    },
...
    "OrderTypeOne": {
      "type": "object",
      "properties": {
        "shipment_type": {
          "type": "number",
          "example": 1
        },
        "order": {
          "type": "object",
          "properties": {
            "comment": {
              "type": "string",
              "example": "blablabla"
            },
            "shippers_shipper_id": {
              "type": "number",
              "example": 3
            },
            "warehouses_warehouse_id": {
              "type": "number",
              "example": 1
            }
          }
        },
        "product": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "product_id": {
                "type": "number",
                "example": 4
              },
              "quantity": {
                "type": "number",
                "example": 2
              }
            }
          }
        }
      }
    },
    "OrderTypeTwo": {
      "type": "object",
      "properties": {
        "shipment_type": {
          "type": "number",
          "example": 2
        },
        "customer": {
          "type": "object",
          "properties": {
            "first_name": {
              "type": "string",
              "example": "David"
            },
            "last_name": {
              "type": "string",
              "example": "Clements"
            },
            "street": {
              "type": "string",
              "example": "59161 Chandler Rapids"
            },
            "email": {
              "type": "string",
              "example": "nguyenpatrick@gmail.com"
            },
            "phone": {
              "type": "string",
              "example": "84081093"
            },
            "cities_postal_code": {
              "type": "number",
              "example": 9000
            },
            "countries_iso": {
              "type": "string",
              "example": "SE"
            },
            "users_user_id": {
              "type": "number",
              "example": 15
            },
            "is_user_profile": {
              "type": "number",
              "example": 0
            }
          }
        },
        "order": {
          "type": "object",
          "properties": {
            "comment": {
              "type": "string",
              "example": "heyheyhey"
            },
            "shippers_shipper_id": {
              "type": "number",
              "example": 2
            },
            "po_boxes_id": {
              "type": "number",
              "example": 3
            },
            "warehouses_warehouse_id": {
              "type": "number",
              "example": 2
            }
          }
        },
        "product": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "product_id": {
                "type": "number",
                "example": 15
              },
              "quantity": {
                "type": "number",
                "example": 8
              }
            }
          }
        }
      }
    },
...
  }
}